# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/journal/entries": {
      "post": {
        "operationId": "JournalReadController_createEntry",
        "summary": "Create a journal entry",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEntryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created entry."
          }
        },
        "tags": [
          "journal"
        ]
      },
      "get": {
        "operationId": "JournalReadController_entries",
        "summary": "List journal entries",
        "description": "Newest first. `from`/`to` bound the instant the entry was lived, as a half-open range [from, to), so an entry exactly at `to` belongs to the next period rather than this one.",
        "parameters": [
          {
            "name": "workspaceId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "ISO 8601",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "ISO 8601",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Entries the caller can see in this workspace."
          }
        },
        "tags": [
          "journal"
        ]
      }
    },
    "/journal/entries/{id}": {
      "patch": {
        "operationId": "JournalReadController_updateEntry",
        "summary": "Edit a journal entry text",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEntryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated entry."
          }
        },
        "tags": [
          "journal"
        ]
      },
      "delete": {
        "operationId": "JournalReadController_deleteEntry",
        "summary": "Delete a journal entry",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Entry deleted."
          }
        },
        "tags": [
          "journal"
        ]
      }
    },
    "/journal/syntheses/{id}": {
      "patch": {
        "operationId": "JournalReadController_updateSynthesis",
        "summary": "Edit a journal synthesis text (never regenerates)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSynthesisDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated synthesis."
          }
        },
        "tags": [
          "journal"
        ]
      }
    },
    "/journal/syntheses": {
      "get": {
        "operationId": "JournalReadController_syntheses",
        "summary": "List journal syntheses",
        "description": "Newest first. Filtered by the start of the period each synthesis covers. There is no `scale` parameter: which scales exist is the Time module's vocabulary, and a synthesis is identified here by its span.",
        "parameters": [
          {
            "name": "workspaceId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "ISO 8601",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "ISO 8601",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Syntheses the caller can see in this workspace."
          }
        },
        "tags": [
          "journal"
        ]
      }
    },
...
    "/references/notes": {
      "post": {
        "operationId": "ReferencesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNoteBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      },
      "get": {
        "operationId": "ReferencesController_list",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "parentId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "take",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skip",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/references/notes/{id}": {
      "get": {
        "operationId": "ReferencesController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "patch": {
        "operationId": "ReferencesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNoteBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "delete": {
        "operationId": "ReferencesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
  "info": {
    "title": "naŭ Platform API",
    "description": "The backend of the naŭ modules.\n\nRoutes under `/v1` are for authenticated people; routes under\n`/_service/v1` are for other naŭ services and are authenticated with a\nservice token instead. The two are separated at the routing level, not\nby a naming convention, so a route cannot end up in the wrong audience\nby having the wrong decorator.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "user-cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      },
      "user-bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "service-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateEntryDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateEntryDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateSynthesisDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateNoteBody": {
        "type": "object",
        "properties": {}
      },
      "UpdateNoteBody": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}